add spring factorybean and example
