+++ examples/src/main/java/io/kubernetes/client/examples/springfactorybean/SpringControllerFactoryBeanExample.java
@@ -10,7 +10,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package io.kubernetes.client.examples;
+package io.kubernetes.client.examples.springfactorybean;
 
 import io.kubernetes.client.extended.controller.Controller;
 import io.kubernetes.client.extended.controller.reconciler.Reconciler;
@@ -27,21 +27,26 @@ import io.kubernetes.client.openapi.models.V1PodList;
 import io.kubernetes.client.spring.extended.controller.KubernetesInformerConfigurer;
 import io.kubernetes.client.spring.extended.controller.KubernetesReconcilerConfigurer;
 import io.kubernetes.client.spring.extended.controller.annotation.*;
+import io.kubernetes.client.spring.extended.controller.factory.KubernetesControllerFactory;
 import io.kubernetes.client.util.ClientBuilder;
 import java.io.IOException;
 import java.time.Duration;
+
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.stereotype.Component;
 
 @SpringBootApplication
-public class SpringControllerExample {
+public class SpringControllerFactoryBeanExample {
 
   public static void main(String[] args) {
-    SpringApplication.run(SpringControllerExample.class, args);
+    SpringApplication.run(SpringControllerFactoryBeanExample.class, args);
   }
 
   @Configuration
@@ -68,11 +73,11 @@ public class SpringControllerExample {
     }
 
     // *REQUIRED*
-    // Configurer components that registers reconciler to the controller-manager in the context.
-    @Bean
-    public KubernetesReconcilerConfigurer kubernetesReconcilerConfigurer(
-        SharedInformerFactory sharedInformerFactory) {
-      return new KubernetesReconcilerConfigurer(sharedInformerFactory);
+    // factorybean to crete controller
+    @Bean("node-printing-controller")
+    public KubernetesControllerFactory kubernetesReconcilerConfigurer(
+        SharedInformerFactory sharedInformerFactory, Reconciler reconciler) {
+      return new KubernetesControllerFactory(sharedInformerFactory, reconciler);
     }
 
     // *OPTIONAL*
@@ -90,12 +95,6 @@ public class SpringControllerExample {
     @Bean("sharedInformerFactory")
     public SharedInformerFactory sharedInformerFactory() {
       return new MySharedInformerFactory();
-    }
-
-    @Bean
-    public NodePrintingReconciler nodePrintingReconciler(
-        Lister<V1Pod> podLister, Lister<V1Node> nodeLister, SharedInformer<V1Node> nodeInformer) {
-      return new NodePrintingReconciler(podLister, nodeLister, nodeInformer);
     }
   }
 
@@ -128,31 +127,35 @@ public class SpringControllerExample {
                 resyncPeriodMillis = 60 * 1000L // fully resync every 1 minute
                 ),
           }))
+  @Component
   public static class NodePrintingReconciler implements Reconciler {
 
-    public NodePrintingReconciler(
-        Lister<V1Pod> podLister, Lister<V1Node> nodeLister, SharedInformer<V1Node> nodeInformer) {
-      this.nodeLister = nodeLister;
-      this.podLister = podLister;
-      this.nodeInformer = nodeInformer;
-    }
-
+    @Value("${namespace}")
+    private String namespace;
+    @Autowired
     private SharedInformer<V1Node> nodeInformer;
-
+    @Autowired
+    private SharedInformer<V1Pod> podInformer;
+    @Autowired
     private Lister<V1Node> nodeLister;
-
+    @Autowired
     private Lister<V1Pod> podLister;
 
     // *OPTIONAL*
     // If you feed like hold the controller from running util some condition..
     @KubernetesReconcilerReadyFunc
-    boolean informerReady() {
-      return nodeInformer.hasSynced();
+    public boolean informerReady() {
+      return podInformer.hasSynced() && nodeInformer.hasSynced();
     }
 
     @Override
     public Result reconcile(Request request) {
       V1Node node = nodeLister.get(request.getName());
+
+      System.out.println("get all pods in namespace "+namespace);
+      podLister.namespace(namespace).list().stream()
+      .map(pod->pod.getMetadata().getName()).forEach(System.out::println);
+      
       System.out.println("triggered reconciling " + node.getMetadata().getName());
       return new Result(false);
     }
