+++ spring/src/main/java/io/kubernetes/client/spring/extended/controller/KubernetesReconcilerProcessor.java
@@ -24,6 +24,8 @@ import io.kubernetes.client.extended.workqueue.RateLimitingQueue;
 import io.kubernetes.client.extended.workqueue.WorkQueue;
 import io.kubernetes.client.informer.SharedInformerFactory;
 import io.kubernetes.client.spring.extended.controller.annotation.*;
+import io.kubernetes.client.spring.extended.controller.factory.KubernetesControllerFactory;
+
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.time.Duration;
@@ -88,232 +90,11 @@ public class KubernetesReconcilerProcessor implements BeanFactoryPostProcessor, 
       KubernetesReconciler kubernetesReconciler =
           reconciler.getClass().getAnnotation(KubernetesReconciler.class);
       String reconcilerName = kubernetesReconciler.value();
-      Controller controller = buildController(this.sharedInformerFactory, reconciler);
+      
+      KubernetesControllerFactory controllerFactory = new KubernetesControllerFactory(sharedInformerFactory, reconciler);
+      
+      Controller controller = controllerFactory.getObject();
       beanFactory.registerSingleton(reconcilerName, controller);
-    }
-  }
-
-  private Controller buildController(SharedInformerFactory sharedInformerFactory, Reconciler r)
-      throws BeansException {
-    KubernetesReconciler kubernetesReconciler =
-        r.getClass().getAnnotation(KubernetesReconciler.class);
-    String reconcilerName = kubernetesReconciler.value();
-
-    KubernetesReconcilerWatches watches = kubernetesReconciler.watches();
-    DefaultControllerBuilder builder = ControllerBuilder.defaultBuilder(sharedInformerFactory);
-    RateLimitingQueue<Request> workQueue = new DefaultRateLimitingQueue<>();
-    builder = builder.withWorkQueue(workQueue);
-    Map<Class, AddFilterAdaptor> addFilters = getAddFilters(watches, r);
-    Map<Class, UpdateFilterAdaptor> updateFilters = getUpdateFilters(watches, r);
-    Map<Class, DeleteFilterAdaptor> deleteFilters = getDeleteFilters(watches, r);
-    List<ReadyFuncAdaptor> readyFuncs = getReadyFuncs(r);
-    for (KubernetesReconcilerWatch watch : watches.value()) {
-      try {
-
-        Function<? extends KubernetesObject, Request> workQueueKeyFunc;
-        try {
-          workQueueKeyFunc =
-              watch.workQueueKeyFunc().getConstructor(WorkQueue.class).newInstance(workQueue);
-        } catch (NoSuchMethodException e) {
-          workQueueKeyFunc = watch.workQueueKeyFunc().newInstance();
-        } catch (InvocationTargetException e) {
-          throw new BeanCreationException(
-              "Failed instantiating controller watch: " + e.getMessage());
-        }
-
-        final Function<? extends KubernetesObject, Request> finalWorkQueueKeyFunc =
-            workQueueKeyFunc;
-        builder =
-            builder.watch(
-                (q) -> {
-                  return ControllerBuilder.controllerWatchBuilder(watch.apiTypeClass(), q)
-                      .withOnAddFilter(addFilters.get(watch.apiTypeClass()))
-                      .withOnUpdateFilter(updateFilters.get(watch.apiTypeClass()))
-                      .withOnDeleteFilter(deleteFilters.get(watch.apiTypeClass()))
-                      .withWorkQueueKeyFunc(finalWorkQueueKeyFunc)
-                      .withResyncPeriod(Duration.ofMillis(watch.resyncPeriodMillis()))
-                      .build();
-                });
-        for (Supplier<Boolean> readyFunc : readyFuncs) {
-          builder = builder.withReadyFunc(readyFunc);
-        }
-      } catch (IllegalAccessException | InstantiationException e) {
-        throw new BeanCreationException("Failed instantiating controller: " + e.getMessage());
-      }
-    }
-
-    builder = builder.withWorkerCount(kubernetesReconciler.workerCount());
-
-    return builder.withReconciler(r).withName(reconcilerName).build();
-  }
-
-  private Map<Class, AddFilterAdaptor> getAddFilters(
-      KubernetesReconcilerWatches watches, Reconciler reconciler) {
-    Map<Class, AddFilterAdaptor> filters = new HashMap<>();
-    Set<Method> allAnnotatedMethods = new HashSet<>();
-    Set<Method> adoptedMethods = new HashSet<>();
-    for (KubernetesReconcilerWatch watch : watches.value()) {
-      for (Method method : reconciler.getClass().getMethods()) {
-        AddWatchEventFilter annotation = method.getAnnotation(AddWatchEventFilter.class);
-        if (annotation != null && watch.apiTypeClass().equals(annotation.apiTypeClass())) {
-          if (filters.containsKey(watch.apiTypeClass())) {
-            log.warn(
-                "Duplicated watch ADD event filter upon apiType {}", annotation.apiTypeClass());
-          }
-          filters.put(watch.apiTypeClass(), new AddFilterAdaptor(reconciler, method));
-          adoptedMethods.add(method);
-        }
-        allAnnotatedMethods.add(method);
-      }
-    }
-    allAnnotatedMethods.removeAll(adoptedMethods);
-    if (allAnnotatedMethods.size() > 0) {
-      log.warn("Dangling watch ADD event filters {}", StringUtils.join(allAnnotatedMethods, ","));
-    }
-    return filters;
-  }
-
-  private Map<Class, UpdateFilterAdaptor> getUpdateFilters(
-      KubernetesReconcilerWatches watches, Reconciler reconciler) {
-    Map<Class, UpdateFilterAdaptor> filters = new HashMap<>();
-    Set<Method> allAnnotatedMethods = new HashSet<>();
-    Set<Method> adoptedMethods = new HashSet<>();
-    for (KubernetesReconcilerWatch watch : watches.value()) {
-      for (Method method : reconciler.getClass().getMethods()) {
-        UpdateWatchEventFilter annotation = method.getAnnotation(UpdateWatchEventFilter.class);
-        if (annotation != null && watch.apiTypeClass().equals(annotation.apiTypeClass())) {
-          if (filters.containsKey(watch.apiTypeClass())) {
-            log.warn(
-                "Duplicated watch UPDATE event filter upon apiType {}", annotation.apiTypeClass());
-          }
-          filters.put(watch.apiTypeClass(), new UpdateFilterAdaptor(reconciler, method));
-          adoptedMethods.add(method);
-        }
-        allAnnotatedMethods.add(method);
-      }
-    }
-    allAnnotatedMethods.removeAll(adoptedMethods);
-    if (allAnnotatedMethods.size() > 0) {
-      log.warn(
-          "Dangling watch UPDATE event filters {}", StringUtils.join(allAnnotatedMethods, ","));
-    }
-    return filters;
-  }
-
-  private Map<Class, DeleteFilterAdaptor> getDeleteFilters(
-      KubernetesReconcilerWatches watches, Reconciler reconciler) {
-    Map<Class, DeleteFilterAdaptor> filters = new HashMap<>();
-    Set<Method> allAnnotatedMethods = new HashSet<>();
-    Set<Method> adoptedMethods = new HashSet<>();
-    for (KubernetesReconcilerWatch watch : watches.value()) {
-      for (Method method : reconciler.getClass().getMethods()) {
-        DeleteWatchEventFilter annotation = method.getAnnotation(DeleteWatchEventFilter.class);
-        if (annotation != null && watch.apiTypeClass().equals(annotation.apiTypeClass())) {
-          if (filters.containsKey(watch.apiTypeClass())) {
-            log.warn(
-                "Duplicated watch DELETE event filter upon apiType {}", annotation.apiTypeClass());
-          }
-          filters.put(watch.apiTypeClass(), new DeleteFilterAdaptor(reconciler, method));
-          adoptedMethods.add(method);
-        }
-        allAnnotatedMethods.add(method);
-      }
-    }
-    allAnnotatedMethods.removeAll(adoptedMethods);
-    if (allAnnotatedMethods.size() > 0) {
-      log.warn(
-          "Dangling watch DELETE event filters {}", StringUtils.join(allAnnotatedMethods, ","));
-    }
-    return filters;
-  }
-
-  private List<ReadyFuncAdaptor> getReadyFuncs(Reconciler reconciler) {
-    List<ReadyFuncAdaptor> readyFuncs = new ArrayList<>();
-    for (Method method : reconciler.getClass().getMethods()) {
-      if (method.isAnnotationPresent(KubernetesReconcilerReadyFunc.class)) {
-        readyFuncs.add(new ReadyFuncAdaptor(reconciler, method));
-      }
-    }
-    return readyFuncs;
-  }
-
-  private static class AddFilterAdaptor implements Predicate {
-    private Method method;
-    private Object target;
-
-    private AddFilterAdaptor(Object target, Method method) {
-      this.method = method;
-      this.target = target;
-    }
-
-    @Override
-    public boolean test(Object addedObj) {
-      try {
-        return (boolean) method.invoke(target, addedObj);
-      } catch (IllegalAccessException | InvocationTargetException e) {
-        log.error("invalid EventAddFilter method signature", e);
-        return true;
-      }
-    }
-  }
-
-  private static class UpdateFilterAdaptor implements BiPredicate {
-    private Method method;
-    private Object target;
-
-    private UpdateFilterAdaptor(Object target, Method method) {
-      this.method = method;
-      this.target = target;
-    }
-
-    @Override
-    public boolean test(Object oldObj, Object newObj) {
-      try {
-        return (boolean) method.invoke(target, oldObj, newObj);
-      } catch (IllegalAccessException | InvocationTargetException e) {
-        log.error("invalid EventUpdateFilter method signature", e);
-        return true;
-      }
-    }
-  }
-
-  private static class DeleteFilterAdaptor implements BiPredicate {
-    private Method method;
-    private Object target;
-
-    private DeleteFilterAdaptor(Object target, Method method) {
-      this.method = method;
-      this.target = target;
-    }
-
-    @Override
-    public boolean test(Object deleteObj, Object cacheStatusUnknown) {
-      try {
-        return (boolean) method.invoke(target, deleteObj, cacheStatusUnknown);
-      } catch (IllegalAccessException | InvocationTargetException e) {
-        log.error("invalid EventDeleteFilter method signature", e);
-        return true;
-      }
-    }
-  }
-
-  private static class ReadyFuncAdaptor implements Supplier<Boolean> {
-    private Method method;
-    private Object target;
-
-    private ReadyFuncAdaptor(Object target, Method method) {
-      this.method = method;
-      this.target = target;
-    }
-
-    @Override
-    public Boolean get() {
-      try {
-        return (boolean) method.invoke(target);
-      } catch (IllegalAccessException | InvocationTargetException e) {
-        log.error("invalid ReadyFunc method signature", e);
-        return false;
-      }
     }
   }
 }
